+++ test/Danom.Tests/Result/ResultTests.cs
@@ -26,7 +26,7 @@ public sealed class ResultTests
     [Fact]
     public void Error()
     {
-        var result = Result<int, string>.Error(new("Error"));
+        var result = Result<int, string>.Error("Error");
         AssertResult.IsError(result);
         Assert.False(result.IsOk);
         Assert.Equal("Error(Error)", result.ToString());
@@ -40,7 +40,7 @@ public sealed class ResultTests
                 .Match(x => x, _ => -1));
 
         Assert.Equal(1,
-            Result<int, string>.Error(new("Error"))
+            Result<int, string>.Error("Error")
                 .Match(_ => -1, _ => 1));
     }
 
@@ -59,7 +59,7 @@ public sealed class ResultTests
         ok = false;
         error = false;
 
-        Result<int, string>.Error(new("Error"))
+        Result<int, string>.Error("Error")
             .Match(ok: _ => ok = true, error: _ => error = true);
 
         Assert.False(ok);
@@ -70,55 +70,55 @@ public sealed class ResultTests
     public void Bind()
     {
         AssertResult.IsOk(2, Result<int, string>.Ok(1).Bind(x => Result<int, string>.Ok(x + 1)));
-        AssertResult.IsError(Result<int, string>.Error(new("Error")).Bind(x => Result<int, string>.Ok(x + 1)));
+        AssertResult.IsError(Result<int, string>.Error("Error").Bind(x => Result<int, string>.Ok(x + 1)));
     }
 
     [Fact]
     public void Map()
     {
         AssertResult.IsOk(2, Result<int, string>.Ok(1).Map(x => x + 1));
-        AssertResult.IsError(Result<int, string>.Error(new("Error")).Map(x => x + 1));
+        AssertResult.IsError(Result<int, string>.Error("Error").Map(x => x + 1));
     }
 
     [Fact]
     public void DefaultValue()
     {
-        Assert.Equal(1, Result<int, string>.Error(new("Error")).DefaultValue(1));
+        Assert.Equal(1, Result<int, string>.Error("Error").DefaultValue(1));
         Assert.Equal(2, Result<int, string>.Ok(2).DefaultValue(1));
     }
 
     [Fact]
     public void DefaultWith()
     {
-        Assert.Equal(1, Result<int, string>.Error(new("Error")).DefaultWith(() => 1));
+        Assert.Equal(1, Result<int, string>.Error("Error").DefaultWith(() => 1));
         Assert.Equal(2, Result<int, string>.Ok(2).DefaultWith(() => 1));
     }
 
     [Fact]
     public void Equality()
     {
-        Assert.Equal(Result<int, string>.Error(new("Error")), Result<int, string>.Error(new("Error")));
+        Assert.Equal(Result<int, string>.Error("Error"), Result<int, string>.Error("Error"));
         Assert.Equal(Result<int, string>.Ok(1), Result<int, string>.Ok(1));
         Assert.NotEqual(Result<int, string>.Ok(1), Result<int, string>.Ok(2));
-        Assert.NotEqual(Result<int, string>.Ok(1), Result<int, string>.Error(new("Error")));
+        Assert.NotEqual(Result<int, string>.Ok(1), Result<int, string>.Error("Error"));
     }
 
     [Fact]
     public void EqualityOperator()
     {
-        Assert.True(Result<int, string>.Error(new("Error")) == Result<int, string>.Error(new("Error")));
+        Assert.True(Result<int, string>.Error("Error") == Result<int, string>.Error("Error"));
         Assert.True(Result<int, string>.Ok(1) == Result<int, string>.Ok(1));
         Assert.False(Result<int, string>.Ok(1) == Result<int, string>.Ok(2));
-        Assert.False(Result<int, string>.Ok(1) == Result<int, string>.Error(new("Error")));
+        Assert.False(Result<int, string>.Ok(1) == Result<int, string>.Error("Error"));
     }
 
     [Fact]
     public void InequalityOperator()
     {
-        Assert.False(Result<int, string>.Error(new("Error")) != Result<int, string>.Error(new("Error")));
+        Assert.False(Result<int, string>.Error("Error") != Result<int, string>.Error("Error"));
         Assert.False(Result<int, string>.Ok(1) != Result<int, string>.Ok(1));
         Assert.True(Result<int, string>.Ok(1) != Result<int, string>.Ok(2));
-        Assert.True(Result<int, string>.Ok(1) != Result<int, string>.Error(new("Error")));
+        Assert.True(Result<int, string>.Ok(1) != Result<int, string>.Error("Error"));
     }
 
     [Fact]
@@ -139,7 +139,7 @@ public sealed class ResultTests
     [Fact]
     public void TryGetError()
     {
-        var resultE = Result<int, string>.Error(new("Error"));
+        var resultE = Result<int, string>.Error("Error");
         if (resultE.TryGet(out var valueE, out var errorE) == false)
         {
             Assert.Equal(default, valueE);
@@ -183,7 +183,7 @@ public sealed class ResultTTests
     [Fact]
     public async Task ErrorAsync()
     {
-        var result = await Task.FromResult(Result<int>.Error(new("Error")));
+        var result = await Task.FromResult(Result<int>.Error("Error"));
         AssertResult.IsError(result);
         Assert.False(result.IsOk);
     }
@@ -191,7 +191,31 @@ public sealed class ResultTTests
     [Fact]
     public void Error()
     {
-        var result = Result<int>.Error(new("Error"));
+        var result = Result<int>.Error(new ResultErrors("Error"));
+        AssertResult.IsError(result);
+        Assert.False(result.IsOk);
+    }
+
+    [Fact]
+    public void ErrorFromString()
+    {
+        var result = Result<int>.Error("Error");
+        AssertResult.IsError(result);
+        Assert.False(result.IsOk);
+    }
+
+    [Fact]
+    public void ErrorFromKeyAndValue()
+    {
+        var result = Result<int>.Error("Error", "Error message");
+        AssertResult.IsError(result);
+        Assert.False(result.IsOk);
+    }
+
+    [Fact]
+    public void ErrorFromKeyAndParams()
+    {
+        var result = Result<int>.Error("Error", "Error message 1", "Error message 2");
         AssertResult.IsError(result);
         Assert.False(result.IsOk);
     }
@@ -204,7 +228,7 @@ public sealed class ResultTTests
                 .Match(x => x, _ => -1));
 
         Assert.Equal(1,
-            Result<int>.Error(new("Error"))
+            Result<int>.Error("Error")
                 .Match(_ => -1, _ => 1));
     }
 
@@ -212,27 +236,27 @@ public sealed class ResultTTests
     public void Bind()
     {
         AssertResult.IsOk(2, Result<int>.Ok(1).Bind(x => Result<int>.Ok(x + 1)));
-        AssertResult.IsError(Result<int>.Error(new("Error")).Bind(x => Result<int>.Ok(x + 1)));
+        AssertResult.IsError(Result<int>.Error("Error").Bind(x => Result<int>.Ok(x + 1)));
     }
 
     [Fact]
     public void Map()
     {
         AssertResult.IsOk(2, Result<int>.Ok(1).Map(x => x + 1));
-        AssertResult.IsError(Result<int>.Error(new("Error")).Map(x => x + 1));
+        AssertResult.IsError(Result<int>.Error("Error").Map(x => x + 1));
     }
 
     [Fact]
     public void DefaultValue()
     {
-        Assert.Equal(1, Result<int>.Error(new("Error")).DefaultValue(1));
+        Assert.Equal(1, Result<int>.Error("Error").DefaultValue(1));
         Assert.Equal(2, Result<int>.Ok(2).DefaultValue(1));
     }
 
     [Fact]
     public void DefaultWith()
     {
-        Assert.Equal(1, Result<int>.Error(new("Error")).DefaultWith(() => 1));
+        Assert.Equal(1, Result<int>.Error("Error").DefaultWith(() => 1));
         Assert.Equal(2, Result<int>.Ok(2).DefaultWith(() => 1));
     }
 }
@@ -245,7 +269,7 @@ public sealed class ResultTAsyncTests
         var resultOk = await Task.FromResult(Result<int>.Ok(1)).MatchAsync(x => x, _ => -1);
         Assert.Equal(1, resultOk);
 
-        var resultError = await Task.FromResult(Result<int>.Error(new("Error"))).MatchAsync(_ => -1, _ => 1);
+        var resultError = await Task.FromResult(Result<int>.Error("Error")).MatchAsync(_ => -1, _ => 1);
         Assert.Equal(1, resultError);
     }
 
@@ -254,8 +278,8 @@ public sealed class ResultTAsyncTests
     {
         AssertResult.IsOk(2, await Task.FromResult(Result<int>.Ok(1)).BindAsync(x => Result<int>.Ok(x + 1)));
         AssertResult.IsOk(2, await Task.FromResult(Result<int>.Ok(1)).BindAsync(x => Result<int>.OkAsync(Task.FromResult(x + 1))));
-        AssertResult.IsError(await Task.FromResult(Result<int>.Error(new("Error"))).BindAsync(x => Result<int>.Ok(x + 1)));
-        AssertResult.IsError(await Task.FromResult(Result<int>.Error(new("Error"))).BindAsync(x => Result<int>.OkAsync(Task.FromResult(x + 1))));
+        AssertResult.IsError(await Task.FromResult(Result<int>.Error("Error")).BindAsync(x => Result<int>.Ok(x + 1)));
+        AssertResult.IsError(await Task.FromResult(Result<int>.Error("Error")).BindAsync(x => Result<int>.OkAsync(Task.FromResult(x + 1))));
     }
 
     [Fact]
@@ -263,15 +287,15 @@ public sealed class ResultTAsyncTests
     {
         AssertResult.IsOk(2, await Task.FromResult(Result<int>.Ok(1)).MapAsync(x => x + 1));
         AssertResult.IsOk(2, await Task.FromResult(Result<int>.Ok(1)).MapAsync(x => Task.FromResult(x + 1)));
-        AssertResult.IsError(await Task.FromResult(Result<int>.Error(new("Error"))).MapAsync(x => x + 1));
-        AssertResult.IsError(await Task.FromResult(Result<int>.Error(new("Error"))).MapAsync(x => Task.FromResult(x + 1)));
+        AssertResult.IsError(await Task.FromResult(Result<int>.Error("Error")).MapAsync(x => x + 1));
+        AssertResult.IsError(await Task.FromResult(Result<int>.Error("Error")).MapAsync(x => Task.FromResult(x + 1)));
     }
 
     [Fact]
     public async Task DefaultValue()
     {
-        Assert.Equal(1, await Task.FromResult(Result<int>.Error(new("Error"))).DefaultValueAsync(1));
-        Assert.Equal(1, await Task.FromResult(Result<int>.Error(new("Error"))).DefaultValueAsync(Task.FromResult(1)));
+        Assert.Equal(1, await Task.FromResult(Result<int>.Error("Error")).DefaultValueAsync(1));
+        Assert.Equal(1, await Task.FromResult(Result<int>.Error("Error")).DefaultValueAsync(Task.FromResult(1)));
         Assert.Equal(2, await Task.FromResult(Result<int>.Ok(2)).DefaultValueAsync(1));
         Assert.Equal(2, await Task.FromResult(Result<int>.Ok(2)).DefaultValueAsync(Task.FromResult(1)));
     }
@@ -279,9 +303,11 @@ public sealed class ResultTAsyncTests
     [Fact]
     public async Task DefaultWith()
     {
-        Assert.Equal(1, await Task.FromResult(Result<int>.Error(new("Error"))).DefaultWithAsync(() => 1));
-        Assert.Equal(1, await Task.FromResult(Result<int>.Error(new("Error"))).DefaultWithAsync(() => Task.FromResult(1)));
+        Assert.Equal(1, await Task.FromResult(Result<int>.Error("Error")).DefaultWithAsync(() => 1));
+        Assert.Equal(1, await Task.FromResult(Result<int>.Error("Error")).DefaultWithAsync(() => Task.FromResult(1)));
         Assert.Equal(2, await Task.FromResult(Result<int>.Ok(2)).DefaultWithAsync(() => 1));
         Assert.Equal(2, await Task.FromResult(Result<int>.Ok(2)).DefaultWithAsync(() => Task.FromResult(1)));
     }
+
+
 }