+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - `DanomPageModel` base class for Razor Pages to provide a consistent API for handling options and results.
+- `Result<T>.Error` static extensions to mirror `ResultErrors` constructor signatures.
 - `Result<T, TError>.TryGet(out T? ok, out TError? error)` to safely provide the internal value. A return value indicates whether or not the Result was Ok or Error.
 - `Option<T>` implementation for `IComparable<T>`.
 - `ResultErrors` keyed error list constructor (ex: `new ResultErrors("key", ["value1", "value2"])`).